Fix even more integer type disagreements
More standardization on the correct `size_t` for when string lengths and
vector offsets are involved.

+++ src/subcommand/layout_main.cpp
@@ -162,7 +162,7 @@ int main_layout(int argc, char **argv) {
     std::function<uint64_t(const std::vector<path_handle_t> &,
                            const xp::XP &)> get_max_path_step_count
         = [&](const std::vector<path_handle_t> &path_sgd_use_paths, const xp::XP &path_index) {
-              uint64_t max_path_step_count = 0;
+              size_t max_path_step_count = 0;
               for (auto& path : path_sgd_use_paths) {
                   max_path_step_count = std::max(max_path_step_count, path_index.get_path_step_count(path));
               }
